ADDING DummyPatients
class that gets list of exaple patients

+++ Examino/Server/Examino.Domain/Entities/Patient.cs
@@ -2,7 +2,7 @@
 {
     public class Patient : ApplicationUser
     {
-        public double? Height { get; set; }
+        public int? Height { get; set; }
         public int? Weight { get; set; }
         public string? BloodType { get; set; }
     }
